rethink the variable assignment

+++ bug-definitions/src/test/java/org/bugby/pattern/example/test/var/VariableAssignCheck2.java
@@ -3,7 +3,7 @@ package org.bugby.pattern.example.test.var;
 public class VariableAssignCheck2 {
 
 	public void method2() {
-		short x = 10;
+		double x = 10;
 		if (x == 2) {
 			return;
 		}
